Purge <stdint.h> and <stdbool.h> from kernel
These do not belong in kernel.

Register and stack values are replaced with unsigned long; fixed-width
struct members are replaced with uxx types (e.g. u16, u32, etc.) or
bit-fields, where appropriate.

+++ include/kernel/kernel.h
@@ -20,9 +20,9 @@
 #define __KERNEL_H_
 
 /* GLOBAL DATA */
-extern unsigned int stack;
-extern unsigned int kstart;
-extern unsigned int kend;
+extern unsigned long stack;
+extern unsigned long kstart;
+extern unsigned long kend;
 extern void(*dfl_sighandlers[])(int);
 
 /* initialization procedures */
